click through progress

+++ csdt/core/blocks.js
@@ -1711,12 +1711,10 @@ export function clear() {
 	this.setVisibility(true);
 	this.hasBorder = false;
 
-	try {
-		this.deleteVariable("base image size");
-		this.deleteVariable("style image size");
-		this.deleteVariable("stylization ratio");
-		this.deleteVariable("conversion mode");
-	} catch (e) {}
+	if (this.isVariableNameInUse("base image size")) this.deleteVariable("base image size");
+	if (this.isVariableNameInUse("style image size")) this.deleteVariable("style image size");
+	if (this.isVariableNameInUse("stylization ratio")) this.deleteVariable("stylization ratio");
+	if (this.isVariableNameInUse("conversion mode")) this.deleteVariable("conversion mode");
 }
 
 export function doSwitchToCostume(id, noShadow) {
